distsql: change logic to set distsql mode
Our previous logic to run a statement using distsql was to use the
default distsql mode if the session's mode was set to off, resulting in
using distsql in cases where the session might have explicitly set it to
off. This change initializes the session with the server's default mode
and uses whatever the session's setting is.
