+++ pkg/sql/executor.go
@@ -42,7 +42,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
 	"github.com/cockroachdb/cockroach/pkg/util"
 	"github.com/cockroachdb/cockroach/pkg/util/duration"
-	"github.com/cockroachdb/cockroach/pkg/util/envutil"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/metric"
@@ -78,52 +77,6 @@ var (
 	MetaMisc               = metric.Metadata{Name: "sql.misc.count"}
 	MetaQuery              = metric.Metadata{Name: "sql.query.count"}
 )
-
-// distSQLExecMode controls if and when the Executor uses DistSQL.
-type distSQLExecMode int
-
-const (
-	// distSQLOff means that we never use distSQL.
-	distSQLOff distSQLExecMode = iota
-	// distSQLAuto means that we automatically decide on a case-by-case basis if
-	// we use distSQL.
-	distSQLAuto
-	// distSQLOn means that we use distSQL for queries that are supported.
-	distSQLOn
-	// distSQLAlways means that we only use distSQL; unsupported queries fail.
-	distSQLAlways
-)
-
-func distSQLExecModeFromString(val string) distSQLExecMode {
-	switch strings.ToUpper(val) {
-	case "OFF":
-		return distSQLOff
-	case "AUTO":
-		return distSQLAuto
-	case "ON":
-		return distSQLOn
-	case "ALWAYS":
-		return distSQLAlways
-	default:
-		panic(fmt.Sprintf("unknown DistSQL mode %s", val))
-	}
-}
-
-// defaultDistSQLMode controls the default DistSQL mode (see above). It can
-// still be overridden per-session using `SET DIST_SQL = ...`.
-var defaultDistSQLMode = distSQLExecModeFromString(
-	envutil.EnvOrDefaultString("COCKROACH_DISTSQL_MODE", "OFF"),
-)
-
-// SetDefaultDistSQLMode changes the default DistSQL mode; returns a function
-// that can be used to restore the previous mode.
-func SetDefaultDistSQLMode(mode string) func() {
-	prevMode := defaultDistSQLMode
-	defaultDistSQLMode = distSQLExecModeFromString(mode)
-	return func() {
-		defaultDistSQLMode = prevMode
-	}
-}
 
 type traceResult struct {
 	tag   string
@@ -1392,11 +1345,7 @@ func (e *Executor) execClassic(planner *planner, plan planNode, result *Result) 
 // shouldUseDistSQL determines whether we should use DistSQL for a plan, based
 // on the session settings.
 func (e *Executor) shouldUseDistSQL(planner *planner, plan planNode) (bool, error) {
-	distSQLMode := defaultDistSQLMode
-	if planner.session.DistSQLMode != distSQLOff {
-		distSQLMode = planner.session.DistSQLMode
-	}
-
+	distSQLMode := planner.session.DistSQLMode
 	if distSQLMode == distSQLOff {
 		return false, nil
 	}
